refactor(cli): split ssh module into file
